Internal restructuring for better order
